Word deletion and better modularization at manage page
