+++ src/src/components/Router.jsx
@@ -5,7 +5,6 @@ import { Routes, Route } from 'react-router-dom';
 import ProtectedRoutes from './ProtectedRoutes';
 
 // Importing all page components
-import HomePage from '../pages/HomePage';
 import LoginPage from '../pages/LoginPage';
 import SignUpPage from '../pages/SignUpPage';
 import NotFoundPage from '../pages/NotFoundPage';
@@ -20,7 +19,7 @@ const AppRoutes = () => {
             <Route path='/'>
 
                 {/* Common pages for visitors */}
-                <Route path='' element={<HomePage />} />
+                <Route path='' element={<LoginPage />} />
                 <Route path='login' element={<LoginPage />} />
                 <Route path='signup' element={<SignUpPage/>}/>
                 <Route path='*' element={<NotFoundPage />}/>
