+++ src/src/components/wordInput.jsx
@@ -1,21 +1,50 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
+import { useGame } from '../contexts/gameContext';
 
-const wordInput = ({ words }) => {
-    [inputStr, setInputStr] = useState('');
+const WordInput = () => {
+    const { addWordTag } = useGame();
+    const [error, setError] = useState('');
+    const [newWord, setNewWord] = useState('');
 
-    useEffect()
-    
+    const handleWordAdding = () => {
+        // Verifying if the giving string is correctly formated
+        const testStr = newWord.trim();
+        if(!testStr.includes(':')) {
+            setError('Invalid format, place the values between \':\'');
+            return;
+        } 
+        
+        // Verifying if it has both words
+        const separatorIndex = testStr.indexOf(':')
+        if(separatorIndex == 0) {
+            setError('You need to inform the original language\'s word');
+            return;
+        }
+        if(separatorIndex == testStr.length - 1) {
+            setError('You need to include at least one translation');
+            return;
+        }
+        
+        // Updating our wordList
+        addWordTag(newWord);
+    }  
 
     return (
-        <>
-        <div className='word-input'>
+        <div className='word-adder'>
+            <label>Add a new word!</label><br/>
             <input 
                 type="text"
-                value={words}
-                onChange={e => setInputStr(e.target.value)}
+                onChange={e => setNewWord(e.target.value)}
+                placeholder='word : translation'
             />
-            <button onClick={handleChange}>Add</button>
+
+            <button 
+                className='word-adder-btn'
+                onClick={handleWordAdding}
+            >add</button>
+            {error && <p className='error-message'>{error}</p>}
         </div>
-        </>
     )
 }
+
+export default WordInput;
