--- src/src/pages/HomePage.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { useState, useEffect, useContext } from 'react';
-
-const HomePage = () => {
-    return (
-        <>
-            <h1>MOCK - home</h1>
-        </>
-    )
-};
-
-export default HomePage;
